feat: 캐러셀 분리
